Split VersionRangeResolveIntegrationTest
Previously the whole class includes 300+ iterations. Now we split it into 3 classes.
