+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/AbstractVersionRangeResolveIntegrationTest.groovy
@@ -21,7 +21,6 @@ import org.gradle.integtests.fixtures.executer.GradleContextualExecuter
 import org.gradle.integtests.fixtures.resolve.ResolveTestFixture
 import org.gradle.resolve.scenarios.VersionRangeResolveTestScenarios
 import spock.lang.IgnoreIf
-import spock.lang.Unroll
 
 /**
  * A comprehensive test of dependency resolution of a single module version, given a set of input selectors.
@@ -33,7 +32,7 @@ import spock.lang.Unroll
     // embedded mode
     !GradleContextualExecuter.embedded
 })
-class VersionRangeResolveIntegrationTest extends AbstractDependencyResolutionTest {
+abstract class AbstractVersionRangeResolveIntegrationTest extends AbstractDependencyResolutionTest {
 
     def baseBuild
     def baseSettings
@@ -58,48 +57,6 @@ class VersionRangeResolveIntegrationTest extends AbstractDependencyResolutionTes
         baseSettings = settingsFile.text
     }
 
-    @Unroll
-    def "resolve pair #permutation"() {
-        given:
-        def candidates = permutation.candidates
-        def expectedSingle = permutation.expectedSingle
-        def expectedMulti = permutation.expectedMulti
-
-        expect:
-        checkScenarioResolution(expectedSingle, expectedMulti, candidates)
-
-        where:
-        permutation << VersionRangeResolveTestScenarios.SCENARIOS_TWO_DEPENDENCIES
-    }
-
-    @Unroll
-    def "resolve prefer pair #permutation"() {
-        given:
-        def candidates = permutation.candidates
-        def expectedSingle = permutation.expectedSingle
-        def expectedMulti = permutation.expectedMulti
-
-        expect:
-        checkScenarioResolution(expectedSingle, expectedMulti, candidates)
-
-        where:
-        permutation << VersionRangeResolveTestScenarios.SCENARIOS_PREFER
-    }
-
-    @Unroll
-    def "resolve reject pair #permutation"() {
-        given:
-        def candidates = permutation.candidates
-        def expectedSingle = permutation.expectedSingle
-        def expectedMulti = permutation.expectedMulti
-
-        expect:
-        checkScenarioResolution(expectedSingle, expectedMulti, candidates)
-
-        where:
-        permutation << VersionRangeResolveTestScenarios.SCENARIOS_DEPENDENCY_WITH_REJECT
-    }
-
     void checkScenarioResolution(String expectedSingle, String expectedMulti, VersionRangeResolveTestScenarios.RenderableVersion... versions) {
         checkScenarioResolution(expectedSingle, expectedMulti, versions as List)
     }
@@ -118,7 +75,7 @@ class VersionRangeResolveIntegrationTest extends AbstractDependencyResolutionTes
             allprojects {
                 configurations { conf }
             }
-            
+
             configurations {
                 ${singleProjectConfs.join('\n')}
                 single {
@@ -131,12 +88,12 @@ class VersionRangeResolveIntegrationTest extends AbstractDependencyResolutionTes
                 conf project(path: ':p1', configuration: 'conf')
                 ${singleProjectDeps.join('\n')}
             }
-            
+
             task resolveMultiProject(type: Sync) {
                 from configurations.conf
                 into 'libs-multi'
             }
-            
+
             task resolveSingleProject(type: Sync) {
                 from configurations.single
                 into 'libs-single'
@@ -178,8 +135,6 @@ class VersionRangeResolveIntegrationTest extends AbstractDependencyResolutionTes
             def singleProjectResolve = file('libs-single').list() as List
             assert parseResolvedVersion(singleProjectResolve) == expectedSingle
         }
-
-
     }
 
     def parseResolvedVersion(resolvedFiles) {
